requirements.txt
pymediainfo
librosa
openai-whisper
matplotlib